garbage_classified_model/requirements.txt
tensorflow
gradio
Pillow
